Example and doc fixups (#35)
* Get example working with version 1.0 of the gem.

This involves updates to work with the new-model stripped-down
'request' Ruby objects, which don't have an embedded appId or
challenges of their own.  (And also updates to the example's
own Gemfile, of course.)

* Update README for version 1.1 U2F/JS API, version 1.0 gem API.

This requires supplying appId and challenge values to JS separately
from the request objects (which no longer have these items bundled),
and deleting 'as_json' method invocations which appear to no longer
be necessary.

* Make README use <%= ... %> ERB syntax consistently.

Corrects some usages of #{...} syntax taken from HAML in the
example code.